Roj-ASM   Fixed some major bugs      - ConstantWriter: direct append constant when #reset      - ConstantPool: buggly compress method   Split into Tree mode and Visitor mode (WIP)   Obfuscator (80%) Roj-Collect   Unioner for long   TimedHashMap update Roj-Concurrent   Concurrented: IntMap, TimedHashMap   Concurrent Maps and TaskPool now using LockSupport to wait   ExecutionTask can wait now Roj-Config   XML Parser     Remove debug data     (Un)Serialize to JSON simplified     Custom Query Language to find entries you desired          module.component[name="NewModuleRootManager"].content          my.data(2).com[type="xyz" | 456, ntr != 233]       Warning: only '=' operator supported now   All parsers:     Support [long], and negative numbers   JSON Parser     ParseMulti, multi json without seperator Roj-IO   MutableZipFile: WIP   XIPFile: WIP Lavac: WIP FMD:   Project, State Cache   Fix some bugs   Use XMLParser to modify IDEA's project paths Roj-Net:   A Simple DNS Server(Only Support UDP)!      Block A Huge Amount of ADs if you like.      Still has some bugs, but you can click F5   HTTP Server Not use NIO anymore, f@@k Roj-Text:   SimpleLineReader: decrease memory Roj-Util:   Helpers.alwaysTrue() Predicate
Updated Readme

+++ roj/math/MathUtils.java
@@ -1000,6 +1000,6 @@ public abstract class MathUtils {
         if (result > 4294967295L || result < Integer.MIN_VALUE)
             throw new NumberFormatException("Value overflow " + result + " : " + s);
 
-        return negative ? (int) -result : (int) result;
+        return (int) (negative ? -result : result);
     }
 }
